requirements.txt
python-instagram
spotipy